+++ backend/zane_api/views/serializers/environments.py
@@ -1,0 +1,15 @@
+import rest_framework.serializers as serializers
+
+
+# ==========================================
+#               Environments               #
+# ==========================================
+
+
+class CreateEnvironmentRequestSerializer(serializers.Serializer):
+    name = serializers.SlugField(max_length=255)
+
+
+class CloneEnvironmentRequestSerializer(serializers.Serializer):
+    deploy_services = serializers.BooleanField(default=False, required=False)
+    name = serializers.SlugField(max_length=255)
