+++ backend/zane_api/views/serializers/services.py
@@ -1,328 +1,47 @@
-import base64
-from io import StringIO
 import json
-import re
-import time
 from typing import Any
 
-import django_filters
-from django.conf import settings
-from django.core.exceptions import ValidationError
-from django.db.models import Q, QuerySet
+from django.db.models import Q
 from django.utils.translation import gettext_lazy as _
-from django_filters import OrderingFilter
-from dotenv import dotenv_values
-from faker import Faker
-from rest_framework import pagination
-from rest_framework.request import Request
 
-from ..dtos import DockerServiceSnapshot, DeploymentChangeDto
-
-from .helpers import (
-    compute_docker_service_snapshot,
+from ..helpers import (
     compute_docker_service_snapshot_with_changes,
     compute_all_deployment_changes,
 )
-from .. import serializers
-from ..models import (
+import rest_framework.serializers as serializers
+from ...models import (
     URL,
-    Deployment,
-    Project,
-    ArchivedProject,
     Service,
     Volume,
     EnvVariable,
     PortConfiguration,
-    HttpLog,
     Config,
-    DeploymentURL,
-    DeploymentChange,
 )
 from temporal.helpers import (
     check_if_docker_image_exists,
     check_if_port_is_available_on_host,
-    get_server_resource_limits,
 )
-from ..utils import (
-    Colors,
+from ...utils import (
     EnhancedJSONEncoder,
-    convert_value_to_bytes,
     find_item_in_sequence,
-    format_storage_value,
 )
-from ..git_client import GitClient
-from ..validators import validate_git_commit_sha, validate_url_path, validate_env_name
+from ...git_client import GitClient
+from ...validators import validate_git_commit_sha
 
-from search.dtos import RuntimeLogLevel, RuntimeLogSource
-from search.serializers import RuntimeLogSerializer
+from .common import (
+    ConfigRequestSerializer,
+    DockerCredentialsRequestSerializer,
+    EnvRequestSerializer,
+    HealthCheckRequestSerializer,
+    ResourceLimitsRequestSerializer,
+    ServicePortsRequestSerializer,
+    URLRequestSerializer,
+    VolumeRequestSerializer,
+)
 
 # ==============================
 #    Docker services create    #
 # ==============================
-
-
-class DockerCredentialsRequestSerializer(serializers.Serializer):
-    username = serializers.CharField(max_length=100, allow_blank=True, required=False)
-    password = serializers.CharField(max_length=100, allow_blank=True, required=False)
-
-    def validate(self, attrs: dict[str, str]):
-        if attrs.get("username") and not attrs.get("password"):
-            raise serializers.ValidationError(
-                {"password": "This field may not be blank."}
-            )
-        elif attrs.get("password") and not attrs.get("username"):
-            raise serializers.ValidationError(
-                {"username": "This field may not be blank."}
-            )
-        return attrs
-
-
-class ServicePortsRequestSerializer(serializers.Serializer):
-    host = serializers.IntegerField(required=True, min_value=1)
-    forwarded = serializers.IntegerField(required=True, min_value=1)
-
-
-class ConfigRequestSerializer(serializers.Serializer):
-    contents = serializers.CharField(required=True, allow_blank=True)
-    name = serializers.CharField(required=False)
-    mount_path = serializers.URLPathField(required=True)
-    language = serializers.CharField(default="plaintext", required=False)
-
-    def validate(self, attrs: dict):
-        if attrs.get("name") is None:
-            fake = Faker()
-            Faker.seed(time.monotonic())
-            attrs["name"] = fake.slug().lower()
-        return attrs
-
-
-class VolumeRequestSerializer(serializers.Serializer):
-    name = serializers.CharField(max_length=255, required=False, min_length=1)
-    container_path = serializers.URLPathField(max_length=255)
-    host_path = serializers.URLPathField(max_length=255, required=False, default=None)
-    VOLUME_MODE_CHOICES = (
-        ("READ_ONLY", _("READ_ONLY")),
-        ("READ_WRITE", _("READ_WRITE")),
-    )
-    mode = serializers.ChoiceField(required=False, choices=VOLUME_MODE_CHOICES)
-
-    def validate(self, attrs: dict):
-        if attrs.get("mode") is None:
-            if attrs.get("host_path") is not None:
-                attrs["mode"] = "READ_ONLY"
-            else:
-                attrs["mode"] = "READ_WRITE"
-        else:
-            if attrs.get("host_path") is not None and attrs.get("mode") != "READ_ONLY":
-                raise serializers.ValidationError(
-                    {
-                        "mode": [
-                            "Volumes with a host path can only be mounted in `read only` mode."
-                        ]
-                    }
-                )
-        if attrs.get("name") is None:
-            fake = Faker()
-            Faker.seed(time.monotonic())
-            attrs["name"] = fake.slug().lower()
-        return attrs
-
-
-class URLRedirectSerializer(serializers.Serializer):
-    url = serializers.URLField(required=True)
-    permanent = serializers.BooleanField(default=False)
-
-
-class URLRequestSerializer(serializers.Serializer):
-    domain = serializers.URLDomainField(required=False)
-    base_path = serializers.URLPathField(required=False, default="/")
-    strip_prefix = serializers.BooleanField(required=False, default=True)
-    redirect_to = URLRedirectSerializer(required=False)
-    associated_port = serializers.IntegerField(required=False, min_value=1)
-
-    def validate(self, attrs: dict):
-        service: Service = self.context.get("service")  # type: ignore
-
-        if attrs.get("domain") is None:
-            attrs["domain"] = URL.generate_default_domain(service)
-
-        if attrs["domain"] == settings.ZANE_APP_DOMAIN:
-            raise serializers.ValidationError(
-                {
-                    "domain": [
-                        "Using the domain where ZaneOps is installed is not allowed."
-                    ]
-                }
-            )
-        if attrs["domain"] == f"*.{settings.ZANE_APP_DOMAIN}":
-            raise serializers.ValidationError(
-                {
-                    "domain": [
-                        "Using the domain where ZaneOps is installed as a wildcard domain is not allowed."
-                    ]
-                }
-            )
-        if attrs["domain"] == f"*.{settings.ROOT_DOMAIN}":
-            raise serializers.ValidationError(
-                {
-                    "domain": [
-                        "Using the root domain as a wildcard is not allowed as it would shadow all the other services installed on ZaneOps."
-                    ]
-                }
-            )
-        existing_urls = URL.objects.filter(
-            Q(domain=attrs["domain"].lower())
-            & Q(base_path=attrs["base_path"].lower())
-            & ~Q(service__id=service.id if service is not None else None)
-        ).distinct()
-        if len(existing_urls) > 0:
-            raise serializers.ValidationError(
-                {
-                    "domain": [
-                        f"URL with domain `{attrs['domain']}` and base path `{attrs['base_path']}` "
-                        f"is already assigned to another service."
-                    ]
-                }
-            )
-
-        existing_deployment_urls = DeploymentURL.objects.filter(
-            Q(domain=attrs["domain"].lower())
-        ).distinct()
-        if len(existing_deployment_urls) > 0:
-            raise serializers.ValidationError(
-                {
-                    "domain": [
-                        f"URL with domain `{attrs['domain']}` is already assigned to another deployment."
-                    ]
-                }
-            )
-
-        domain = attrs["domain"]
-        domain_parts = domain.split(".")
-        domain_as_wildcard = domain.replace(domain_parts[0], "*", 1)
-
-        existing_parent_domain = URL.objects.filter(
-            Q(domain=domain_as_wildcard.lower())
-            & ~Q(service=service)
-            & Q(base_path=attrs["base_path"].lower())
-        ).distinct()
-        if len(existing_parent_domain) > 0:
-            raise serializers.ValidationError(
-                {
-                    "domain": [
-                        f"URL with domain `{attrs['domain']}` cannot be used because it will be shadowed by the wildcard"
-                        f" domain `{domain_as_wildcard}` which is already assigned to another service."
-                    ]
-                }
-            )
-
-        if attrs.get("associated_port") is None and attrs.get("redirect_to") is None:
-            raise serializers.ValidationError(
-                {
-                    "associated_port": [
-                        "To expose this service, you need to add an associated port to forward this URL to."
-                    ]
-                }
-            )
-        elif (
-            attrs.get("associated_port") is not None
-            and attrs.get("redirect_to") is not None
-        ):
-            raise serializers.ValidationError(
-                {
-                    "associated_port": [
-                        "You cannot provide an associated port if this URL is redirect URL."
-                    ]
-                }
-            )
-
-        return attrs
-
-
-class MemoryLimitRequestSerializer(serializers.Serializer):
-    MEMORY_UNITS = (
-        ("BYTES", _("bytes")),
-        ("KILOBYTES", _("kilobytes")),
-        ("MEGABYTES", _("megabytes")),
-        ("GIGABYTES", _("gigabytes")),
-    )
-    value = serializers.IntegerField(min_value=0)
-    unit = serializers.ChoiceField(choices=MEMORY_UNITS, default="MEGABYTES")
-
-    def validate(self, attrs: dict[str, int | str]):
-        _, max_memory = get_server_resource_limits()
-        six_megabytes = 6 * 1024 * 1024
-        value_in_bytes = convert_value_to_bytes(int(attrs["value"]), attrs["unit"])  # type: ignore
-        # The documentation for docker says that we can't use less than 6mb of memory :
-        # https://docs.docker.com/engine/containers/resource_constraints/#limit-a-containers-access-to-memory
-        if value_in_bytes < six_megabytes:
-            raise serializers.ValidationError(
-                {"value": "Cannot limit a service memory to less than 6 MiB."}
-            )
-        if value_in_bytes > max_memory:
-            raise serializers.ValidationError(
-                {
-                    "value": f"The maximum memory limit on this server is {format_storage_value(max_memory)}."
-                }
-            )
-
-        return attrs
-
-
-class ResourceLimitsRequestSerializer(serializers.Serializer):
-    cpus = serializers.FloatField(required=False, min_value=0.1)
-    memory = MemoryLimitRequestSerializer(required=False)
-
-    def validate(self, attrs: dict):
-        max_cpus, _ = get_server_resource_limits()
-
-        cpu_limit = attrs.get("cpus", 0)
-        if cpu_limit > max_cpus:
-            raise serializers.ValidationError(
-                {"cpus": f"Cannot exceed {max_cpus} CPUs."}
-            )
-
-        return attrs
-
-
-class HealthCheckRequestSerializer(serializers.Serializer):
-    HEALTHCHECK_CHOICES = (
-        ("PATH", _("path")),
-        ("COMMAND", _("command")),
-    )
-    type = serializers.CustomChoiceField(required=True, choices=HEALTHCHECK_CHOICES)
-    value = serializers.CharField(max_length=255, required=True)
-    timeout_seconds = serializers.IntegerField(required=False, default=30, min_value=5)
-    interval_seconds = serializers.IntegerField(required=False, default=30, min_value=5)
-    associated_port = serializers.IntegerField(required=False, min_value=1)
-
-    def validate(self, attrs: dict):
-        if attrs["type"] == "PATH":
-            try:
-                validate_url_path(attrs["value"])
-            except ValidationError as e:
-                raise serializers.ValidationError({"value": e.messages})
-
-            if attrs.get("associated_port") is None:
-                raise serializers.ValidationError(
-                    {
-                        "associated_port": "This field is required.",
-                    }
-                )
-        else:
-            if attrs.get("associated_port") is not None:
-                raise serializers.ValidationError(
-                    {
-                        "associated_port": "Cannot specify an associated port for healthcheck of types `COMMAND`.",
-                    }
-                )
-
-        return attrs
-
-
-class EnvRequestSerializer(serializers.Serializer):
-    key = serializers.CharField(required=True, validators=[validate_env_name])
-    value = serializers.CharField(required=True, allow_blank=True)
 
 
 class DockerServiceCreateRequestSerializer(serializers.Serializer):
@@ -500,115 +219,6 @@ class GitServiceWebhookDeployRequestSerializer(serializers.Serializer):
     commit_sha = serializers.CharField(
         default="HEAD", validators=[validate_git_commit_sha]
     )
-
-
-# ==============================
-#       Docker deployments     #
-# ==============================
-
-
-class DockerServiceDeploymentFilterSet(django_filters.FilterSet):
-    status = django_filters.MultipleChoiceFilter(
-        choices=Deployment.DeploymentStatus.choices
-    )
-    queued_at = django_filters.DateTimeFromToRangeFilter()
-
-    class Meta:
-        model = Deployment
-        fields = ["status", "queued_at"]
-
-
-class DeploymentListPagination(pagination.PageNumberPagination):
-    page_size = 10
-    page_size_query_param = "per_page"
-    page_query_param = "page"
-
-
-# ==============================
-#        Projects List         #
-# ==============================
-
-
-class ProjectListFilterSet(django_filters.FilterSet):
-    sort_by = OrderingFilter(
-        fields=["slug", "updated_at"],
-        field_labels={
-            "slug": "name",
-        },
-    )
-    slug = django_filters.CharFilter(lookup_expr="icontains")
-
-    class Meta:
-        model = Project
-        fields = ["slug"]
-
-
-class ArchivedProjectListFilterSet(django_filters.FilterSet):
-    sort_by = OrderingFilter(
-        fields=["slug", "archived_at"],
-        field_labels={
-            "slug": "name",
-        },
-    )
-    slug = django_filters.CharFilter(lookup_expr="icontains")
-
-    class Meta:
-        model = ArchivedProject
-        fields = ["slug"]
-
-
-class ProjectListPagination(pagination.PageNumberPagination):
-    page_size = 10
-    page_size_query_param = "per_page"
-    page_query_param = "page"
-
-
-# ==============================
-#       Projects Create        #
-# ==============================
-
-
-class ProjectCreateRequestSerializer(serializers.Serializer):
-    slug = serializers.SlugField(max_length=255, required=False)
-    description = serializers.CharField(required=False)
-
-
-# ==============================
-#       Projects Update        #
-# ==============================
-
-
-class ProjectUpdateRequestSerializer(serializers.Serializer):
-    slug = serializers.SlugField(max_length=255, required=False)
-    description = serializers.CharField(required=False)
-
-    def validate(self, attrs: dict[str, str]):
-        if not bool(attrs):
-            raise serializers.ValidationError(
-                "one of `slug` or `description` should be provided"
-            )
-        return attrs
-
-
-# ==============================
-#       Projects Search        #
-# ==============================
-
-
-class ProjectSearchSerializer(serializers.Serializer):
-    id = serializers.CharField(required=True)
-    created_at = serializers.DateTimeField(required=True)
-    slug = serializers.SlugField(required=True)
-    type = serializers.ChoiceField(choices=["project"], default="project")
-
-
-class ServiceSearchSerializer(serializers.Serializer):
-    id = serializers.CharField(required=True)
-    project_slug = serializers.SlugField(required=True)
-    slug = serializers.SlugField(required=True)
-    created_at = serializers.DateTimeField(required=True)
-    type = serializers.ChoiceField(choices=["service"], default="service")
-    environment = serializers.CharField(required=True)
 
 
 # ==============================
@@ -888,7 +498,7 @@ class VolumeItemChangeSerializer(BaseChangeItemSerializer):
                 raise serializers.ValidationError(
                     {
                         "new_value": {
-                            "host_path": f"Cannot remove the host path from a volume that was originally mounted with one, "
+                            "host_path": "Cannot remove the host path from a volume that was originally mounted with one, "
                             f"you need to delete and recreate the volume without the host path."
                         }
                     }
@@ -901,7 +511,7 @@ class VolumeItemChangeSerializer(BaseChangeItemSerializer):
                 raise serializers.ValidationError(
                     {
                         "new_value": {
-                            "host_path": f"Cannot mount a volume to a host path if it wasn't originally mounted that way, "
+                            "host_path": "Cannot mount a volume to a host path if it wasn't originally mounted that way, "
                             f"you need to delete and recreate the volume with a host path."
                         }
                     }
@@ -1225,436 +835,6 @@ class DockerDeploymentFieldChangeRequestSerializer(serializers.Serializer):
             "configs",
         ],
     )
-
-
-# ==============================
-#       Collect Logs           #
-# ==============================
-
-
-class DockerContainerLogSerializer(serializers.Serializer):
-    log = serializers.CharField(required=True, allow_blank=True, trim_whitespace=False)
-    container_id = serializers.CharField(required=True)
-    container_name = serializers.CharField(required=True)
-    time = serializers.CharField(required=True)
-    tag = serializers.CharField(required=True)
-    SOURCES = (
-        ("stdout", _("standard ouput")),
-        ("stderr", _("standard error")),
-    )
-    source = serializers.ChoiceField(choices=SOURCES, required=True)
-
-
-class HTTPServiceRequestSerializer(serializers.Serializer):
-    remote_ip = serializers.IPAddressField(required=True)
-    client_ip = serializers.IPAddressField(required=True)
-    remote_port = serializers.CharField(required=True)
-    PROTOCOLS = [
-        ("HTTP/1.0", "HTTP/1.0"),
-        ("HTTP/1.1", "HTTP/1.1"),
-        ("HTTP/2.0", "HTTP/2.0"),
-        ("HTTP/3.0", "HTTP/3.0"),
-    ]
-    REQUEST_METHODS = [
-        ("GET", "GET"),
-        ("POST", "POST"),
-        ("PUT", "PUT"),
-        ("DELETE", "DELETE"),
-        ("PATCH", "PATCH"),
-        ("OPTIONS", "OPTIONS"),
-        ("HEAD", "HEAD"),
-    ]
-    proto = serializers.ChoiceField(choices=PROTOCOLS, required=True)
-    method = serializers.ChoiceField(choices=REQUEST_METHODS, required=True)
-    host = serializers.CharField(required=True)
-    uri = serializers.CharField(required=True)
-    headers = serializers.DictField(
-        child=serializers.ListField(child=serializers.CharField()),
-        required=True,
-    )
-
-
-class HTTPServiceLogSerializer(serializers.Serializer):
-    ts = serializers.FloatField(required=True)
-    msg = serializers.CharField(required=True)
-    LOG_LEVELS = (
-        ("debug", _("debug")),
-        ("info", _("info")),
-        ("warn", _("warn")),
-        ("error", _("error")),
-        ("panic", _("panic")),
-        ("fatal", _("fatal")),
-    )
-    level = serializers.ChoiceField(choices=LOG_LEVELS, required=True)
-
-    duration = serializers.FloatField()
-    status = serializers.IntegerField(min_value=100)
-    resp_headers = serializers.DictField(
-        child=serializers.ListField(child=serializers.CharField())
-    )
-    request = HTTPServiceRequestSerializer()
-    zane_deployment_upstream = serializers.CharField()
-    zane_deployment_green_hash = serializers.CharField(
-        allow_null=True, required=False, allow_blank=True
-    )
-    zane_deployment_blue_hash = serializers.CharField(
-        allow_null=True, required=False, allow_blank=True
-    )
-    zane_service_id = serializers.CharField()
-    zane_deployment_id = serializers.CharField(required=False)
-    uuid = serializers.CharField(allow_null=True, required=False, allow_blank=True)
-
-
-class DockerContainerLogsRequestSerializer(serializers.ListSerializer):
-    child = DockerContainerLogSerializer()
-
-
-class DockerContainerLogsResponseSerializer(serializers.Serializer):
-    simple_logs_inserted = serializers.IntegerField(min_value=0)
-    http_logs_inserted = serializers.IntegerField(min_value=0)
-
-
-# =======================================
-#        Deployment BUILD Logs        #
-# =======================================
-
-
-class DeploymentBuildLogsQuerySerializer(serializers.Serializer):
-    cursor = serializers.CharField(required=False)
-    per_page = serializers.IntegerField(
-        required=False, min_value=1, max_value=100, default=50
-    )
-
-    def validate_cursor(self, cursor: str):
-        try:
-            decoded_data = base64.b64decode(cursor, validate=True)
-            decoded_string = decoded_data.decode("utf-8")
-            serializer = CursorSerializer(data=json.loads(decoded_string))
-            serializer.is_valid(raise_exception=True)
-        except (serializers.ValidationError, ValueError):
-            raise serializers.ValidationError(
-                {
-                    "cursor": "Invalid cursor format, it should be a base64 encoded string of a JSON object."
-                }
-            )
-        return cursor
-
-
-# =======================================
-#        Deployment runtime Logs        #
-# =======================================
-
-
-class DeploymentRuntimeLogsQuerySerializer(serializers.Serializer):
-    time_before = serializers.DateTimeField(required=False)
-    time_after = serializers.DateTimeField(required=False)
-    query = serializers.CharField(
-        required=False, allow_blank=True, trim_whitespace=False
-    )
-    level = serializers.ListField(
-        child=serializers.ChoiceField(
-            choices=[RuntimeLogLevel.INFO, RuntimeLogLevel.ERROR]
-        ),
-        required=False,
-    )
-    per_page = serializers.IntegerField(
-        required=False, min_value=1, max_value=100, default=50
-    )
-    cursor = serializers.CharField(required=False)
-
-    def validate_cursor(self, cursor: str):
-        try:
-            decoded_data = base64.b64decode(cursor, validate=True)
-            decoded_string = decoded_data.decode("utf-8")
-            serializer = CursorSerializer(data=json.loads(decoded_string))
-            serializer.is_valid(raise_exception=True)
-        except (serializers.ValidationError, ValueError):
-            raise serializers.ValidationError(
-                {
-                    "cursor": "Invalid cursor format, it should be a base64 encoded string of a JSON object."
-                }
-            )
-        return cursor
-
-
-class CursorSerializer(serializers.Serializer):
-    sort = serializers.ListField(required=True, child=serializers.CharField())
-    order = serializers.ChoiceField(choices=["desc", "asc"], required=True)
-
-
-class DeploymentHttpLogsPagination(pagination.CursorPagination):
-    page_size = 50
-    page_size_query_param = "per_page"
-    ordering = ("-time",)
-
-    def get_ordering(self, request: Request, queryset, view):
-        filter = DeploymentHttpLogsFilterSet(
-            {"sort_by": ",".join(request.GET.getlist("sort_by"))}
-        )
-
-        if filter.is_valid():
-            ordering = tuple(
-                set(filter.form.cleaned_data.get("sort_by", self.ordering))
-            )
-            if len(ordering) > 0:
-                return ordering  # tuple(set(filter.form.cleaned_data.get("sort_by", self.ordering)))
-
-        return self.ordering
-
-
-class DeploymentHttpLogsFilterSet(django_filters.FilterSet):
-    time = django_filters.DateTimeFromToRangeFilter()
-    request_method = django_filters.MultipleChoiceFilter(
-        choices=HttpLog.RequestMethod.choices
-    )
-    sort_by = OrderingFilter(fields=["time", "request_duration_ns"])
-    request_query = django_filters.CharFilter(
-        field_name="request_query", method="filter_query"
-    )
-    status = django_filters.BaseInFilter(method="filter_multiple_values")
-    request_ip = django_filters.BaseInFilter(method="filter_multiple_values")
-    request_user_agent = django_filters.BaseInFilter(method="filter_multiple_values")
-    request_host = django_filters.BaseInFilter(
-        field_name="request_host", method="filter_multiple_values"
-    )
-    request_path = django_filters.BaseInFilter(method="filter_multiple_values")
-
-    def filter_multiple_values(self, queryset: QuerySet, name: str, value: str):
-        params = self.request.GET.getlist(name)  # type: ignore
-
-        status_prefix_path = r"^\dxx$"
-
-        queries = Q()
-        if name == "status":
-            for param in params:
-                if re.match(status_prefix_path, param):
-                    prefix = int(param[0])
-                    queries = queries | (
-                        Q(status__gte=(prefix * 100), status__lte=(prefix * 100) + 99)
-                    )
-                elif re.match(r"^\d+$", param):
-                    queries = queries | Q(status=int(param))
-        else:
-            queries = Q(**{f"{name}__in": params})
-        print(f"Query: {Colors.GREY}{queries}{Colors.ENDC}")
-        return queryset.filter(queries)
-
-    def filter_query(self, queryset: QuerySet, name: str, value: str):
-        return queryset.filter(request_query__istartswith=value)
-
-    class Meta:
-        model = HttpLog
-        fields = [
-            "time",
-            "request_method",
-            "request_path",
-            "request_host",
-            "request_query",
-            "status",
-            "request_ip",
-            "request_id",
-            "request_user_agent",
-        ]
-
-
-# ==============================
-#       Http logs fields       #
-# ==============================
-
-
-class HttpLogFieldsQuerySerializer(serializers.Serializer):
-    field = serializers.ChoiceField(
-        choices=[
-            "request_host",
-            "request_path",
-            "request_user_agent",
-            "request_ip",
-        ]
-    )
-    value = serializers.CharField(allow_blank=True)
-
-
-class HttpLogFieldsResponseSerializer(serializers.ListSerializer):
-    child = serializers.CharField()
-
-
-# ==============================
-#     Project Service List     #
-# ==============================
-
-
-class ServiceListParamSerializer(serializers.Serializer):
-    query = serializers.CharField(required=False)
-
-
-class BaseServiceCardSerializer(serializers.Serializer):
-    updated_at = serializers.DateTimeField(required=True)
-    volume_number = serializers.IntegerField(required=True)
-    slug = serializers.CharField(required=True)
-    url = serializers.URLField(allow_null=True)
-    STATUS_CHOICES = (
-        ("HEALTHY", _("Healthy")),
-        ("UNHEALTHY", _("Unhealthy")),
-        ("FAILED", _("Failed")),
-        ("SLEEPING", _("Sleeping")),
-        ("NOT_DEPLOYED_YET", _("Not deployed yet")),
-        ("DEPLOYING", _("Deploying")),
-    )
-    status = serializers.ChoiceField(choices=STATUS_CHOICES)
-    id = serializers.CharField(required=True)
-
-
-class DockerServiceCardSerializer(BaseServiceCardSerializer):
-    type = serializers.ChoiceField(choices=["docker"], default="docker")
-    image = serializers.CharField(required=True)
-    tag = serializers.CharField(required=True)
-
-
-class GitServiceCardSerializer(BaseServiceCardSerializer):
-    type = serializers.ChoiceField(choices=["git"], default="git")
-    repository = serializers.CharField(required=True)
-    last_commit_message = serializers.CharField(required=False)
-    branch = serializers.CharField(required=True)
-
-
-# ==============================
-#       Resources Search       #
-# ==============================
-
-
-class ResourceSearchParamSerializer(serializers.Serializer):
-    query = serializers.CharField(required=False)
-
-
-# =============================================
-#       Load env variables as big string      #
-# =============================================
-
-
-class EnvStringChangeSerializer(serializers.Serializer):
-    new_value = serializers.CharField(required=True, allow_blank=True)
-
-    def validate(self, attrs: dict):
-        service: Service | None = self.context.get("service")
-        if service is None:
-            raise serializers.ValidationError("`service` is required in context.")
-
-        envs = dotenv_values(stream=StringIO(attrs["new_value"]))
-        errors = []
-        for key, value in envs.items():
-            try:
-                validate_env_name(key)
-            except ValidationError as err:
-                errors.append(f"`{key}` is not a valid variable name : {err.message}")
-
-            if value is None:
-                envs[key] = ""
-        if len(errors) > 0:
-            raise serializers.ValidationError({"new_value": errors})
-
-        new_value = ""
-        for key, value in envs.items():
-            new_value += f"{key}='{value}'\n"
-        attrs["new_value"] = new_value
-
-        # validate double `key`
-        env_changes = [
-            DeploymentChangeDto(
-                type="ADD",
-                field=DeploymentChange.ChangeField.ENV_VARIABLES,
-                new_value={
-                    "key": key,
-                    "value": value,
-                },
-            )
-            for key, value in envs.items()
-        ]
-        snapshot = compute_docker_service_snapshot(
-            DockerServiceSnapshot.from_dict(
-                serializers.ServiceSerializer(service).data  # type: ignore
-            ),
-            [*env_changes, *compute_all_deployment_changes(service)],
-        )
-
-        for env in snapshot.duplicate_envs:
-            errors.append(f"variable with name `{env}` already exists in the service")
-
-        if len(errors) > 0:
-            raise serializers.ValidationError({"new_value": errors})
-
-        return attrs
-
-
-# ==========================================
-#       Service & deployment metrics       #
-# ==========================================
-
-
-class ServiceMetricsSerializer(serializers.Serializer):
-    bucket_epoch = serializers.DateTimeField()
-    avg_cpu = serializers.FloatField()
-    avg_memory = serializers.FloatField()
-    total_net_tx = serializers.IntegerField()
-    total_net_rx = serializers.IntegerField()
-    total_disk_read = serializers.IntegerField()
-    total_disk_write = serializers.IntegerField()
-
-
-class ServiceMetricsResponseSerializer(serializers.ListSerializer):
-    child = ServiceMetricsSerializer()
-
-
-class ServiceMetricsQuery(serializers.Serializer):
-    time_range = serializers.ChoiceField(
-        choices=["LAST_HOUR", "LAST_6HOURS", "LAST_DAY", "LAST_WEEK", "LAST_MONTH"],
-        required=False,
-        default="LAST_HOUR",
-    )
-
-
-# ==========================================
-#              User Creation               #
-# ==========================================
-
-
-class UserExistenceResponseSerializer(serializers.Serializer):
-    exists = serializers.BooleanField()
-
-
-class UserCreationRequestSerializer(serializers.Serializer):
-    username = serializers.CharField(min_length=1, max_length=255)
-    password = serializers.CharField(min_length=8)
-
-
-class UserCreatedResponseSerializer(serializers.Serializer):
-    detail = serializers.CharField()
-
-
-# ==========================================
-#       AUTO UPDATE DOCKER SERVICES        #
-# ==========================================
-
-
-class AutoUpdateRequestSerializer(serializers.Serializer):
-    desired_version = serializers.CharField(required=True, max_length=255)
-
-
-class AutoUpdateResponseSerializer(serializers.Serializer):
-    message = serializers.CharField()
-
-
-# ==========================================
-#               Environments               #
-# ==========================================
-
-
-class CreateEnvironmentRequestSerializer(serializers.Serializer):
-    name = serializers.SlugField(max_length=255)
-
-
-class CloneEnvironmentRequestSerializer(serializers.Serializer):
-    deploy_services = serializers.BooleanField(default=False, required=False)
-    name = serializers.SlugField(max_length=255)
 
 
 # ==========================================
